apps/hyperparam-tuning/requirements.txt
pandas
scikit-learn
pyyaml
category_encoders